+++ api-server/src/main/java/com/ThreeK_Project/api_server/domain/order/controller/OrderAdminController.java
@@ -1,0 +1,28 @@
+package com.ThreeK_Project.api_server.domain.order.controller;
+
+import com.ThreeK_Project.api_server.domain.order.dto.RequestDto.OrderSearchDTO;
+import com.ThreeK_Project.api_server.domain.order.service.OrderService;
+import com.ThreeK_Project.api_server.global.dto.SuccessResponse;
+import com.ThreeK_Project.api_server.global.security.auth.UserDetailsCustom;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.UUID;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/admin/orders")
+public class OrderAdminController {
+
+    private final OrderService orderService;
+
+    @DeleteMapping("/{orderId}")
+    public ResponseEntity<SuccessResponse> deleteOrder(@PathVariable("orderId") UUID orderId) {
+        UserDetailsCustom userDetails = (UserDetailsCustom) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        orderService.deleteOrder(orderId, userDetails.getUser());
+        return ResponseEntity.ok(new SuccessResponse("주문 삭제 성공"));
+    }
+
+}
