+++ api-server/src/test/java/com/ThreeK_Project/api_server/domain/order/controller/OrderAdminControllerTest.java
@@ -1,0 +1,59 @@
+package com.ThreeK_Project.api_server.domain.order.controller;
+
+import com.ThreeK_Project.api_server.customMockUser.WithCustomMockUser;
+import com.ThreeK_Project.api_server.domain.order.service.OrderService;
+import com.ThreeK_Project.api_server.domain.payment.service.PaymentService;
+import com.ThreeK_Project.api_server.domain.user.entity.User;
+import com.ThreeK_Project.api_server.global.security.auth.UserDetailsCustom;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+
+import java.util.UUID;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.doNothing;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@ExtendWith(SpringExtension.class)
+class OrderAdminControllerTest {
+    @Mock
+    private OrderService orderService;
+
+    @InjectMocks
+    private OrderAdminController orderAdminController;
+
+    private MockMvc mockMvc;
+
+    @BeforeEach
+    public void init() {
+        mockMvc = MockMvcBuilders.standaloneSetup(orderAdminController).build();
+    }
+
+    @Test
+    @DisplayName("주문 삭제 성공 테스트")
+    @WithCustomMockUser
+    public void deleteOrder() throws Exception {
+        UUID orderId = UUID.randomUUID();
+        UserDetailsCustom userDetails = (UserDetailsCustom) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User user = userDetails.getUser();
+
+        doNothing().when(orderService).deleteOrder(orderId, user);
+
+        mockMvc.perform(delete("/api/admin/orders/" + orderId))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("message").value("주문 삭제 성공"));
+    }
+
+}
